Update Welcome.tsx
the original message was confusing and grammatically incorrect. changing it to "Every address can claim a maximum of 300,000 CKB per calendar month." effectively conveys the message that for each specific address or user, there is a limit of 300,000 CKB that they can claim within a single calendar month. This claimable amount resets at the start of each new month, allowing users to claim up to 300,000 CKB again in the following month. 

Signed-off-by: anni <112425356+anipenguin@users.noreply.github.com>